put all classes into Duke package
