+++ src/duke/java/Command.java
@@ -1,4 +1,4 @@
-package main.java;
+package duke.java;
 
 public abstract class Command {
     public abstract void execute(TaskList tasks,Ui ui, Storage storage);
