+++ src/duke/java/DeleteCommand.java
@@ -1,19 +1,17 @@
-package main.java;
+package duke.java;
+
 import java.io.IOException;
-import java.util.ArrayList;
 
-public class MarkCommand extends Command {
+public class DeleteCommand extends Command {
 
-    private int taskToMark;
+    int taskToDelete;
 
-    public MarkCommand(int taskToMark) {
-        this.taskToMark = taskToMark;
+    public DeleteCommand(int taskToDelete) {
+        this.taskToDelete = taskToDelete;
     }
-
-    @Override
     public void execute(TaskList tasks,Ui ui, Storage storage) {
-        tasks.mark(taskToMark);
-        ui.showMarkTask(tasks.getTask(taskToMark));
+        ui.showRemoveTask(tasks.getTask(taskToDelete), tasks.size() - 1);
+        tasks.delete(taskToDelete);
         try {
             storage.writeToFile(tasks.getTaskArr());
         } catch (IOException e) {
