+++ src/duke/java/Parser.java
@@ -1,4 +1,4 @@
-package main.java;
+package duke.java;
 
 public class Parser {
