+++ src/duke/java/Storage.java
@@ -1,4 +1,4 @@
-package main.java;
+package duke.java;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
