+++ src/duke/java/Ui.java
@@ -1,4 +1,4 @@
-package main.java;
+package duke.java;
 import java.util.ArrayList;
 import java.util.Scanner;
